move input_poll_overlay to input_driver.c
