+++ input/input_types.h
@@ -17,6 +17,8 @@
 #ifndef __INPUT_TYPES__H
 #define __INPUT_TYPES__H
 
+#include "../msg_hash.h"
+
 enum input_auto_game_focus_type
 {
    AUTO_GAME_FOCUS_OFF = 0,
@@ -25,11 +27,36 @@ enum input_auto_game_focus_type
    AUTO_GAME_FOCUS_LAST
 };
 
-typedef struct rarch_joypad_driver input_device_driver_t;
-typedef struct input_keyboard_line input_keyboard_line_t;
-typedef struct rarch_joypad_info rarch_joypad_info_t;
-typedef struct input_driver input_driver_t;
-typedef struct input_keyboard_ctx_wait input_keyboard_ctx_wait_t;
+struct retro_keybind
+{
+   /* Human-readable label for the control. */
+   char     *joykey_label;
+   /* Human-readable label for an analog axis. */
+   char     *joyaxis_label;
+   /*
+    * Joypad axis. Negative and positive axes are both 
+    * represented by this variable.
+    */
+   uint32_t joyaxis;
+   /* Default joy axis binding value for resetting bind to default. */
+   uint32_t def_joyaxis;
+   /* Used by input_{push,pop}_analog_dpad(). */
+   uint32_t orig_joyaxis;
+
+   enum msg_hash_enums enum_idx;
+
+   enum retro_key key;
+
+   uint16_t id;
+   /* What mouse button ID has been mapped to this control. */
+   uint16_t mbutton;
+   /* Joypad key. Joypad POV (hats) are embedded into this key as well. */
+   uint16_t joykey;
+   /* Default key binding value (for resetting bind). */
+   uint16_t def_joykey;
+   /* Determines whether or not the binding is valid. */
+   bool valid;
+};
 
 typedef struct
 {
@@ -37,6 +64,12 @@ typedef struct
    uint16_t analogs[8];
    uint16_t analog_buttons[16];
 } input_bits_t;
+
+typedef struct rarch_joypad_driver input_device_driver_t;
+typedef struct input_keyboard_line input_keyboard_line_t;
+typedef struct rarch_joypad_info rarch_joypad_info_t;
+typedef struct input_driver input_driver_t;
+typedef struct input_keyboard_ctx_wait input_keyboard_ctx_wait_t;
 
 typedef struct joypad_connection joypad_connection_t;
 typedef struct pad_connection_listener_interface pad_connection_listener_t;
